requirements/ci.txt
-r base.txt

coverage==3.6
django-discover-runner==0.2.2
django-jenkins==0.13.0
